Separate consumers to individual files and update missing Rogue Consumer reference from router.
