+++ setup.py
@@ -5,7 +5,7 @@ with open('requirements.txt') as f:
 
 setup(
     name="quasar",
-    version="1.8.0",
+    version="1.9.0",
     packages=find_packages(),
     install_requires=requirements,
     entry_points={
